Add initial support for TWW compartment clicks
The addon compartment is using the new menu system in 11.0 and has a few
changes that we need to paper over with respect to getting the button
name through the system.

At the moment tooltips don't work, and neither do non-left clicks. These
are both bugs in the compartment that Blizzard needs to resolve.